Add PostgreSQL service to GitHub Actions workflow
The modification includes the addition of a PostgreSQL service to the GitHub Actions workflow in the scala.yml file. This will allow the service to run on the latest Ubuntu and set specific environmental variables for the database. The 'Run tests' step has also been updated to accommodate new database configurations.